+++ service/controller/clusterapi/v29/resource/tcdp/template/params_main.go
@@ -1,6 +1,12 @@
 package template
 
-// ParamsMain is the data structure for the Control Plane Initializer template.
+// ParamsMain is the data structure for the Tenant Cluster Data Plane template.
 type ParamsMain struct {
-	IAMRoles *ParamsMainIAMRoles
+	AutoScalingGroup    *ParamsMainAutoScalingGroup
+	IAMPolicies         *ParamsMainIAMPolicies
+	LaunchConfiguration *ParamsMainLaunchConfiguration
+	LifecycleHooks      *ParamsMainLifecycleHooks
+	Outputs             *ParamsMainOutputs
+	SecurityGroups      *ParamsMainSecurityGroups
+	Subnets             *ParamsMainSubnets
 }
