--- service/controller/clusterapi/v29/resource/tcdp/template/params_main_iam_roles.go
@@ -1,18 +1,0 @@
-package template
-
-type ParamsMainIAMRoles struct {
-	PeerAccessRoleName string
-	Tenant             ParamsMainIAMRolesTenant
-}
-
-type ParamsMainIAMRolesTenant struct {
-	AWS ParamsMainIAMRolesTenantAWS
-}
-
-type ParamsMainIAMRolesTenantAWS struct {
-	Account ParamsMainIAMRolesTenantAWSAccount
-}
-
-type ParamsMainIAMRolesTenantAWSAccount struct {
-	ID string
-}
