+++ service/controller/clusterapi/v29/resource/tcdp/template/template_main_auto_scaling_group.go
