+++ service/controller/clusterapi/v29/resource/tcdp/template/template_main_launch_configuration.go
