+++ service/controller/clusterapi/v29/resource/tcdp/template/template_main_lifecycle_hooks.go
