--- tcdp/error.go
@@ -1,72 +1,0 @@
-package tcdp
-
-import (
-	"strings"
-
-	"github.com/aws/aws-sdk-go/service/cloudformation"
-	"github.com/giantswarm/microerror"
-)
-
-var deleteInProgressError = &microerror.Error{
-	Kind: "deleteInProgressError",
-}
-
-// IsDeleteInProgress asserts deleteInProgressError.
-func IsDeleteInProgress(err error) bool {
-	c := microerror.Cause(err)
-
-	if c == nil {
-		return false
-	}
-
-	if strings.Contains(c.Error(), cloudformation.ResourceStatusDeleteInProgress) {
-		return true
-	}
-
-	if c == deleteInProgressError {
-		return true
-	}
-
-	return false
-}
-
-var executionFailedError = &microerror.Error{
-	Kind: "executionFailedError",
-}
-
-// IsExecutionFailed asserts executionFailedError.
-func IsExecutionFailed(err error) bool {
-	return microerror.Cause(err) == executionFailedError
-}
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
-
-var notExistsError = &microerror.Error{
-	Kind: "notExistsError",
-}
-
-// IsNotExists asserts notExistsError.
-func IsNotExists(err error) bool {
-	c := microerror.Cause(err)
-
-	if c == nil {
-		return false
-	}
-
-	if strings.Contains(c.Error(), "does not exist") {
-		return true
-	}
-
-	if c == notExistsError {
-		return true
-	}
-
-	return false
-}
